✨ Move poll comments into a sheet opened from a floating trigger
Replaces the inline discussion card on the invite and poll admin pages
with a sheet opened from a floating bottom-right button showing the
comment count. Comments are deprecated (being replaced by response
notes + host broadcast), so the trigger deliberately lives outside the
page layout: the event card and voting interface can now be redesigned
without comments in the way, and removal later is just deleting the
overlay mount and the component.

The sheet keeps the existing list and permissions, with a reworked
composer: a single input group (name row for guests, comment textarea,
Enter to submit) that locks after one comment per page load to add
friction against comment spam.
