+++ apps/web/public/locales/en/app.json
@@ -140,6 +140,9 @@
   "commentsSettingLegacyBadge": "Legacy",
   "commentsSettingPhaseOutHint": "Comments are being phased out. Participants can include a note with their response instead.",
   "commentsSettingTitle": "Comments",
+  "commentsSheetDescription": "View and add comments on this poll",
+  "commentsSheetEmpty": "No comments yet",
+  "commentsSheetSubmitted": "Your comment has been added",
   "configuredByEnvironmentVariable": "This setting has been configured by environment variable.",
   "confirm": "Confirm",
   "confirmSpaceName": "Please type the space name to confirm:",
