+++ apps/web/src/app/[locale]/invite/[urlId]/invite-page.tsx
@@ -4,7 +4,7 @@ import { Alert, AlertAction, AlertDescription } from "@rallly/ui/alert";
 import { ArrowUpRightIcon, CrownIcon } from "lucide-react";
 import Link from "next/link";
 import { usePoll } from "@/features/poll/client";
-import Discussion from "@/features/poll/components/discussion";
+import { CommentsSheet } from "@/features/poll/components/comments-sheet";
 import { EventCard } from "@/features/poll/components/event-card";
 import { PollFooter } from "@/features/poll/components/poll-footer";
 import { ResponsiveResults } from "@/features/poll/components/responsive-results";
@@ -58,8 +58,11 @@ export function InvitePage() {
         <VotingForm>
           <ResponsiveResults />
         </VotingForm>
-        <Discussion />
         <PollFooter />
+        <div className="fixed right-4 bottom-4 z-40 lg:right-6 lg:bottom-6">
+          <CommentsSheet className="rounded-full shadow-lg" />
+        </div>
+        <div className="h-12 lg:hidden" />
       </main>
     </div>
   );
